Fix the form submission logic (to some extent)
There used to be two submission buttons: an invisible one inside the
form and a visible one outside. The latter was not attached to the
form in any way, but had an event handler attached. This explains why
pressing the return key while the query field was active never did
what you expected. This has now been solved by unifying these buttons
and by attaching the event handler to the form submission rather than
to the button click.

Needless to say, the entire thing could use some cleaning-up.

Fixes #15.